requirements.txt
Flask==3.1.2
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
pandas==2.3.3
python-dotenv==1.2.1
SQLAlchemy==2.0.45
Werkzeug==3.1.5
google-generativeai
